pyproject.toml
[tool.poetry]
name = "multi-agent"
version = "0.1.0"
description = "A multi-agent system project"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "core" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
# Add your project dependencies here
# For example:
# requests = "^2.31.0"
# pandas = "^2.0.0"
google-genai = "^1.10.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
